Replace `OutputStream` with `ByteBuffer` and introduce `SceneryManager` for chunk registration
- Add toggleable debug logs for elapsed cycle profiling based on topic; https://rune-server.org/threads/server-lag-question.704323/post-5788881
- Rename ByteBufferExtensions
- Post-conversion test corrections
- Add more Synth debug names from https://rune-server.org/threads/synth-debug-names-4982-4997.698908/
- replace incorrect static call to Location.getDistance with instance method call

d

d

d

+++ Emulator/src/main/content/region/asgarnia/burthope/guild/warriors_guild/CatapultRoomPlugin.kt
@@ -117,7 +117,7 @@ class CatapultRoomPlugin :
     }
 
     override fun enter(e: Entity): Boolean {
-        if (e is Player && e.getLocation().z == 1) {
+        if (e is Player && e.getLocation().z.toInt() == 1) {
             players.add(e)
             if (!pulse.isRunning) {
                 pulse.restart()
